GH-1245 Move net_logger_impl to net_logger.hpp
